🐛 FIX: Improved project directory structure
