+++ src/components/courses/index.jsx
@@ -1,0 +1,4 @@
+export { default as CourseList } from './CourseList'
+export { default as Filters } from './Filters'
+export { default as PageNo } from './PageNo'
+export { default as Searchbar } from './Searchbar'
